requirements.txt
absl-py==2.1.0
astunparse==1.6.3
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.0.4
cycler==0.12.1
dominate==2.7.0
flatbuffers==1.12
gast==0.4.0
google-auth==2.37.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.69.0
h5py==3.12.1
idna==3.10
imageio==2.9.0
importlib_metadata==8.5.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
kornia==0.6.8
libclang==18.1.1
lpips==0.1.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.4.2
mdurl==0.1.2
ml_dtypes==0.5.0
mtcnn==0.1.1
namex==0.0.8
networkx==3.2.1
ninja==1.11.1
numpy==1.22.4
oauthlib==3.2.2
opencv-python==4.6.0.66
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
Pillow==9.3.0
protobuf==3.19.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.19.0
pyparsing==3.2.1
pyspng==0.1.1
python-dateutil==2.9.0.post0
PyWavelets==1.6.0
PyYAML==5.3
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
scikit-image==0.19.3
scipy==1.7.1
six==1.17.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.2
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tifffile==2024.8.30
torch==1.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.62.2
trimesh==3.16.2
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.0
zipp==3.21.0
--extra-index-url https://download.pytorch.org/whl/cu113
# neural_renderer_pytorch==1.1.3
# nvdiffrast @ file:///3DpSp/dataset_preprocessing/ffhq/Deep3DFaceRecon_pytorch/nvdiffrast